Fixes for sphinxext
Restore check that restricts the use of FunctionDocumenter
to nodes under a ServiceDocumenter to prevent it from being
used for classes not related to the web service.

Replace direct uses of __name__ with datatypename().

Fix a problem that showed up when get_doc() does not find
any docstrings for a function.

Use for:else instead of checking a sentinal variable after
a search loop.