Add samples_slot option to TypeDocumenter
Allow the output of the TypeDocumenter to come in different
orders, depending on the sample-slot option. Default to
showing samples after the rendered docstring, but also
allow it to come before the docstring by setting the
option to 'before-docstring'. Use 'none' to disable the
sample output for a type entirely.