Sort set in type exception
The values are stored in a set, therefore they are not sorted when used
in the exception message. Later, the unit test might fail because it
checks for the exact message expected, and that may depends on the order
Python decided to return the set. Let's sort it each time to be sure the
message we can expect.

Change-Id: I9e9c1270ea408f9e217320e30902cb74e690fce8