XML Support not required for Ironic
It's not possible to change the set of content_types
that are rendered in WSME. This means that projects
such as ironic have to support all the content_types
that WSME provides, whether that is appropriate or not.

Ironic makes use of HTTP PATCH, in a way that is not
supportable in XML using current tooling.  This patch
allows WSME exposed methods to declare what content
types are to be supported.

Change-Id: If7675e4d08e881e90f984bcd6fcd02b423ad3f7f
Closes-bug: #1271317