Added MultiType class to wsme.types
There are some usecases in Ironic and Tuskar where we the value type
could be either an int or string. Usually, this happens when we validate
a dictionary - it's normal situation, that we have different types for
dict values.

Ironic have it's own workaround for this situation - see
https://github.com/openstack/ironic/blob/master/ironic/api/controllers/v1/types.py#L153

So there is a sense to handle this situation in WSME.

Co-authored-by: Yuriy Zveryanskyy <yzveryanskyy@mirantis.com>
Change-Id: I386ce82c51473a82b9975dec1e31993cb79b3930