Pecan: Make it possible to use the Response to return a non-default return type
This patch is extending the Response object and the Pecan extension
to make it possible to return a non-default return type (just like it
already supports retuning a non default status code).

This is useful when the return code of a function is 204 (No Response)
meaning that the response body message is empty and the content-lentgh
is 0.

Change-Id: I47fdb743a36dd502e8805cf0ee61a89cd922d97a