WIP: REST protocol accept handling allow failover
The changes in 8710dabb652dae775dee31789e91608f832e62e6 broke
protocol failover when the REST protocol is listed before others
(see bug referenced below).

This patch attempts to remedy that problem but it is challenging to
figure out a way to preserve the original semantics of Protcol.accept()
"am I willing to try to handle this protocol" while also enabling the
desired "raise a 406 or 415 when content negotiation fails".

This patch adds (and makes pass) the test in the referenced bug and
leaves to fail those test which represent the content negotiation
use case.

The attempted (but failing) change is to revert to the old mechanics
in accept() and do the content negotiation in iter_calls. This
doesn't work for some cases, though, as iter_calls is never reached
because accept() has already rejected protocol selection.

As far as I can tell there's no easy way to go deeper into protocol
handling before saying "I don't want to handle this request" and
move on to the next request. If accept() returns True then no other
protocol handler will get a chance.

Further: It seems that the (lame) workaround I suggest in the bug
report: Put the REST protocol at the end of the protcol stack may be
the way to go.

Ideas?

Related-Bug: #1442710
Change-Id: I328a392151013c46207519c245213d5dec48ecc9