Clean up setup.py and add requirements.txt
Previously setup.py had information about which versions of libraries to depend
on based on the python version. Since we don't need to support those versions
anymore (2.5, 2.6 and 3.3), these dependancies can be listed in a
requirements.txt file.  This can be set up to sync with global-requirements at
a later date.

Change-Id: I90c4cbd7961ffdf6769578986c272bf9f5721d77