[WIP] Enable real testing of python 3.4
This started as an effort to get testing of python 3.4 actually
testing something and ended up being a bit more (to get things
working at all):

* suds is replaced by the fork called suds-jurko (which appears to
  have the most use)
* targets for not py27 and not py34 are removed

This will fail, as it turns out the py3* tests haven't really been
doing much for quite some time. Most of the errors appear to be
bytes v str related, which will be teased out. I'm pushing this up
to compare my results with the gate.

Change-Id: I837c249714fd957790ea84aa2fd9ad994a39c5ea