Fix the setup.cfg metadata
This patch is removing the quotes from "author" and "author-email"
metadata in the setup.cfg file. Apparently having the values quoted
causes the openstack-infra/project-config release.sh script to error out
when uploading it to Pypi.

This patch also updates the "url" metadata to point to the openstack git
server.

Change-Id: I992a5053a48f7a49f08f299b170cc3da40de242b