Improve handling of :module: in wsme documenters
It appears that in newer sphinx versions the index of :module:
directive in the resulting documenter object array has changed,
it is not the last element anymore. Apart from that, it is
possible to have several :module: entries in class for example.

This change fixes the issue for service, function and type
wsme documenters.

Change-Id: Icf44f8af0c1e6e70d8e921615efe03b845885be7